First version that actually does anything.
It can connect to IRC, be invited to channels and know to join them. If
kicked it knows to stay out. A user can start a game but no one else
can join yet and nothing actually happens.

The main bit of work is in the database structure and the basics of
issuing an IRC WHOIS command in order to find out if someone is
identified to a registered nickname.