Black Card should be stored as index on Game.
The game's current Black Card was being stored in every UserGame object,
which is totally pointless as there's only going to be one BCard per
game at any given time.

There was some confusion with the bcards table, which is storing the
entire deck of Black Cards for this game (and all other games).

This commit alters the Black Card handling to be like White Card
handling: For White Cards the current hand is taken off of the deck and
kept in the user's hand (UserGameHand). So equally the current Black
Card should be taken off of the deck and kept with the game (Game
object).