Display a header/footer around plays. Fixes #12.
User complained that the list of plays displayed when all plays are in
was too difficult to visually separate from the rest of the channel
text. A suggestion was to wrap the list in a header of '=' characters.

This change uses a header the same length as the:

    "nickname: Which is the best play?"

line. e.g:

    <AgainstDev> grifferz: Which is the best play?
    <AgainstDev> =================================