Don't deal new Black Card on resume. Fixes #19.
A new Black Card was being dealt even for games which already had one
and were just being resumed.

This required a schema change as games.bcardidx was INTEGER NOT NULL;
its default value was 0, which is a valid value for an index into the
Black Card deck. So, there was no value it could have which would
indicate it was unset.

This change allows bcardidx to be NULL, and that is what it will be on
game creation. A new Black Card will only be dealt if bcardidx is NULL.