Need to be able to NULL card indices. Issue #75.
When changing packs, the cards in players' hands may need fixing up. If
by chance the new index of a card is the same as the old index of an
existing card for some other player then there would be a constraint
violation, so now we set all card indices for the hand cards to NULL
first. That requires the column to be nullable.