Reclaim our nickname when necessary. Fixes #94.
Respond to the 433 (Nickname in use) numeric by issuing a GHOST command
to NickServ.

Also check that we have our correct nickname:

    - As soon as we see the End of MOTD numeric.

    - Every 5 minutes.

If we don't, then attempt to change to it. This will result in a 433
numeric if it's in use by someone else.

If NickServ tells us that an interloper has been GHOSTed, try to change
to our nick again.