Don't reset activity timer when prepping plays.
Fixes #126.

Move the code that resets the game's activity timer outside of
prep_plays().

prep_plays() is called both when the round is first completed, and any
time the round is completed due to player resignation. It's undesirable
to potentially extend the round time out by another full turnclock just
because someone resigned.