Store order in which packs are added
Issue #140.

At several points we are using the packs() and pack_descs() methods to
construct a list of packs in use, but since this module began to use a
hash instead of an array it has lost the implicit ordering of packs.

This is causing spurious rebuilds of the deck, since the caller thinks
that the current deck is different to the configured deck (different
order of packs).

This change stores the configuration order in the hash for later use.