topup_hands() needs to be after pick_new_tsar()
Players who are waiting to join the game are added during the
pick_new_tsar() code path. topup_hands() was moved to be after
pick_new_tsar() in every code path except the one where the Tsar is
resigned, where it was forgotten.

So in the case where the game was progressed by a Tsar resigning, any
waiting players would be joined with an empty hand.

Fixes #143.