Message absent Tsar to tell them to pick winner
When the round ends and it's time for the Card Tsar to pick a winner,
the plays are simply listed into the channel with the Tsar's nickname
used to force a highlight.

If the Tsar isn't in the channel, however, they will never know that
they need to do something.

This change checks to see if the Card Tsar is in the channel, and sends
them a message if not. It also explaisn this to the channel.

Issue #136.