Do something useful with IRC "No such nick" msgs
If we're sending messages to a nickname that isn't on the network any
more than this is something that the games the user is in should know
about. The other players may be the last chance to get the user back
online.

So, when we receive a 401 numeric ("No such nick"), work out which
active games they are in and tell those games.

Fixes #136.