Fix placeholders with too few underscores
Placeholders need to have at least 5 underscores. Without that, the bot
assumes the underscores are part of the text and that the black card
should accept a single white card at the end.

Perhaps the bot should warn if it sees more than 1 but fewer than 5
underscores in a row?